🐛 fix(investTableData.ts): change parsing of cotacao from parseInt to parseFloat to ensure accurate calculation of currentValue ✨ feat(request.ts): add setData function to update data in Firestore and cache it in Redis 🐛 fix(bonds.ts): add missing import statement for IBondItem interface ✨ feat(bonds.ts): add getBonds function to retrieve all bond items from Firestore ✨ feat(international.ts): add getInternationalAssets function to retrieve all international assets from Firestore ✨ feat(user/walletRecommendation.ts): add userRecommendationUpdate function to update user stocks and reits based on common keys with all stocks and reits ✨ feat(routes.ts): add routes for getting bonds, international assets, and syncing user recommendation update ✨ feat(services/bonds.ts): add getBonds function to handle HTTP request for retrieving all bond items ✨ feat(services/international.ts): add getInternationalAssets function to handle HTTP request for retrieving all international assets ✨ feat(services/user.ts): add userRecommendationUpdate function to handle HTTP request for syncing user recommendation update

+++ src/builders/investTableData.ts
@@ -53,7 +53,7 @@ export const buildAssetTableData = ({
     asset: assets[item].nome,
     recommended: recommendedPercentages[item].percentage,
     currentValue:
-      parseInt(assets[item].quantity) * parseInt(assets[item].cotacao),
+      parseInt(assets[item].quantity) * parseFloat(assets[item].cotacao),
     recommendedValue: 0,
     adjustment: '',
     grade: assetPoints[item],
